documentation and re-organizing emotes data class
